preparing server setup, tsc done
