+++ server/__tests__/server.test.ts
@@ -1,0 +1,2 @@
+import supertest from "supertest";
+import { app } from "../server";
